Prefer infix operator at end of line
If the second argument of an infix doesn't fit on the current line but
fits on a new line, prefer a newline before the argument over partially
formatting it on the same line as the operator.

This gives a preference to having the operator alone on the end of the
line and nicely lines up parenthesized expressions in general.